refactor(test): extract afterSpawnCalled helper to helpers.ts
Move the deterministic spawn polling helper from health.test.ts to the
shared helpers.ts file so it can be reused across all test files.

Part of #29
